DB funcional do projeto , com atualizações no banco

+++ db/migrate/20191018014756_create_usuarios.rb
@@ -1,13 +1,12 @@
 class CreateUsuarios < ActiveRecord::Migration[5.1]
   def change
     create_table :usuarios do |t|
-      t.int :cpf
+      t.bigint :cpf
       t.string :nome
       t.string :senha
       t.string :email
-      t.int :telefone
+      t.bigint :telefone
       t.boolean :capacidadeDeGerencia
-
       t.timestamps
     end
   end
